Perbaikan halaman progress

+++ database/migrations/2025_11_17_122903_add_halaman_to_setoran_table.php
@@ -24,14 +24,21 @@ return new class extends Migration
     }
 
     public function down()
-    {
-        Schema::table('setoran', function (Blueprint $table) {
-            if (!Schema::hasColumn('setoran', 'halaman')) {
-                $table->string('halaman')->nullable();
-            }
+{
+    Schema::table('setoran', function (Blueprint $table) {
 
+        if (Schema::hasColumn('setoran', 'halaman_awal')) {
             $table->dropColumn('halaman_awal');
+        }
+
+        if (Schema::hasColumn('setoran', 'halaman_akhir')) {
             $table->dropColumn('halaman_akhir');
-        });
-    }
+        }
+
+        if (!Schema::hasColumn('setoran', 'halaman')) {
+            $table->string('halaman')->nullable();
+        }
+    });
+}
+
 };
